[modularize] Put files back because we have not decide the path of modules.
